requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
wordcloud
Pillow
tensorflow
